[ui] Create search filter server-side
