+++ app/src/app/search/hooks/use-filter.ts
@@ -1,8 +1,7 @@
-import {Tables} from "@/lib/database.types";
 import {useReducer} from "react";
-import type {SearchFilter, SearchFilterActions, SearchFilterCondition} from "@/app/search/search.types";
+import type {FilterOptions, SearchFilter, SearchFilterActions} from "@/app/search/search.types";
 import {useSearchParams} from "next/navigation";
-import {PHYSICAL_REGION_PATH, PRIMARY_ACTIVITY_CATEGORY_PATH} from "@/app/search/constants";
+import {createFilters} from "@/app/search/create-filters";
 
 function searchFilterReducer(state: SearchFilter[], action: SearchFilterActions): SearchFilter[] {
   switch (action.type) {
@@ -38,102 +37,8 @@ function searchFilterReducer(state: SearchFilter[], action: SearchFilterActions)
   }
 }
 
-interface FilterOptions {
-  activityCategories: Tables<"activity_category_available">[],
-  regions: Tables<"region_used">[]
-  statisticalVariables: Tables<"stat_definition">[]
-}
-
-export const useFilter = ({regions = [], activityCategories = [], statisticalVariables = []}: FilterOptions) => {
-  const urlSearchParams = useSearchParams()
-  const standardFilters: SearchFilter[] = [
-    {
-      type: "search",
-      label: "Name",
-      // Search is a vector field of name indexed for fast full text search.
-      name: "search",
-      selected: urlSearchParams?.has('search') ? [urlSearchParams?.get('search') as string] : [],
-      postgrestQuery: ({selected}) => generateFTSQuery(selected[0])
-    },
-    {
-      type: "search",
-      label: "Tax ID",
-      name: "tax_reg_ident",
-      selected: urlSearchParams?.has('tax_reg_ident') ? [urlSearchParams?.get('tax_reg_ident') as string] : [],
-      postgrestQuery: ({selected}) => selected[0] ? `eq.${selected[0]}` : null
-    },
-    {
-      type: "options",
-      label: "Type",
-      name: "unit_type",
-      options: [
-        {
-          label: "Legal Unit",
-          value: "legal_unit",
-          humanReadableValue: "Legal Unit",
-          className: "bg-legal_unit-200"
-        },
-        {
-          label: "Establishment",
-          value: "establishment",
-          humanReadableValue: "Establishment",
-          className: "bg-establishment-200"
-        },
-        {
-          label: "Enterprise",
-          value: "enterprise",
-          humanReadableValue: "Enterprise",
-          className: "bg-enterprise-200"
-        }
-      ],
-      selected: urlSearchParams?.get('unit_type')?.split(',') ?? ["enterprise"],
-      postgrestQuery: ({selected}) => selected.length ? `in.(${selected.join(',')})` : null
-    },
-    {
-      type: "radio",
-      name: "physical_region_path",
-      label: "Region",
-      options: regions.map(({code, path, name}) => (
-        {
-          label: `${code} ${name}`,
-          value: path as string,
-          humanReadableValue: `${code} ${name}`
-        }
-      )),
-      selected: urlSearchParams?.has(PHYSICAL_REGION_PATH) ? [urlSearchParams?.get(PHYSICAL_REGION_PATH) as string] : [],
-      postgrestQuery: ({selected}) => selected.length ? `cd.${selected.join()}` : null
-    },
-    {
-      type: "radio",
-      name: "primary_activity_category_path",
-      label: "Activity Category",
-      options: activityCategories.map(({code, path, name}) => (
-        {
-          label: `${code} ${name}`,
-          value: path as string,
-          humanReadableValue: `${code} ${name}`
-        }
-      )),
-      selected: urlSearchParams?.has(PRIMARY_ACTIVITY_CATEGORY_PATH) ? [urlSearchParams?.get(PRIMARY_ACTIVITY_CATEGORY_PATH) as string] : [],
-      postgrestQuery: ({selected}) => selected.length ? `cd.${selected.join()}` : null
-    }
-  ];
-
-  const statisticalVariableFilters: SearchFilter[] = statisticalVariables.map(variable => {
-    const conditionalSearchParam = urlSearchParams?.get(variable.code)
-    const [condition, value] = conditionalSearchParam?.split('.') ?? []
-
-    return {
-      type: "conditional",
-      name: variable.code,
-      label: variable.name,
-      selected: value ? [value] : [],
-      condition: condition ? condition as SearchFilterCondition : undefined,
-      postgrestQuery: ({condition, selected}: SearchFilter) =>
-        condition && selected.length ? `${condition}.${selected.join(',')}` : null
-    }
-  });
-
+export const useFilter = (filters: [SearchFilter[], SearchFilter[]]) => {
+  const [standardFilters, statisticalVariableFilters] = filters
   return useReducer(searchFilterReducer, [...standardFilters, ...statisticalVariableFilters])
 }
