+++ app/src/app/search/page.tsx
@@ -2,11 +2,13 @@ import {createClient} from "@/lib/supabase/server";
 import Search from "@/app/search/components/search";
 import {Metadata} from "next";
 
+import {createFilters} from "@/app/search/create-filters";
+
 export const metadata: Metadata = {
     title: "StatBus | Search statistical units"
 }
 
-export default async function SearchPage() {
+export default async function SearchPage({ searchParams }: { readonly searchParams: URLSearchParams }) {
     const client = createClient();
 
     const regionPromise = client
@@ -40,6 +42,12 @@ export default async function SearchPage() {
         console.error('⚠️failed to fetch activity categories', activityCategoriesError);
     }
 
+    const filters = createFilters({
+        activityCategories: activityCategories ?? [],
+        regions: regions ?? [],
+        statisticalVariables: statisticalVariables ?? []
+    }, new URLSearchParams(searchParams));
+
     return (
         <main className="flex flex-col py-8 px-2 md:py-24 mx-auto w-full max-w-5xl">
             <h1 className="font-medium text-xl text-center mb-12">Search for statistical units</h1>
@@ -47,6 +55,7 @@ export default async function SearchPage() {
                 regions={regions ?? []}
                 activityCategories={activityCategories ?? []}
                 statisticalVariables={statisticalVariables ?? []}
+                filters={filters}
             />
         </main>
     )
